+++ src/Base/Creators/RepositoryCreator.php
@@ -1,30 +1,12 @@
 <?php
 
-namespace Inz\Base;
+namespace Inz\Base\Creators;
 
 use Illuminate\Support\Arr;
+use Inz\Base\Abstractions\Creator;
 
-class ContractCreator extends Creator
+class RepositoryCreator extends Creator
 {
-    /**
-     * Model name inserted by the developer.
-     *
-     * @var String
-     */
-    private $modelName;
-    /**
-     * Stub path of the file that will be generated.
-     *
-     * @var String
-     */
-    private $stub;
-    /**
-     * Key value pairs array of parts that will be replaced in the content.
-     *
-     * @var array
-     */
-    private $replacements = [];
-
     public function __construct(String $input)
     {
         parent::__construct();
@@ -33,56 +15,59 @@ class ContractCreator extends Creator
         if (Arr::has($values, 'subdirectory')) {
             $this->subdirectory = $values['subdirectory'];
         }
-        $this->stub = __DIR__ . '/Stubs/Contracts/ExampleRepository.stub';
-        $this->classNameSuffix = 'RepositoryInterface';
-        $this->initializeReplacementsParts($this->modelName);
+        $this->stub = __DIR__ . '/../Stubs/Eloquent/EloquentExampleRepository.stub';
+        $this->classNameSuffix = 'Repository';
     }
 
     /**
      * Initialize the array of the parts that will be replaced.
      *
-     * @param String $modelName
+     * @param String $contractNamespace
+     * @param String $contractName
+     * @param String $modelNamespace
      * @return void
      */
-    public function initializeReplacementsParts(String $modelName)
+    public function initializeReplacementsParts(String $contractNamespace, String $contractName, String $modelNamespace)
     {
         $this->replacements = [
-            '%namespaces.contracts%' => $this->baseNamespace . $this->namespaceConfig . $this->subdirectory,
-            '%modelName%' => $modelName,
+            '%contract%' => $contractNamespace,
+            '%contractName%' => $contractName,
+            '%model%' => $modelNamespace,
+            '%modelName%' => $this->modelName,
+            '%namespaces.repositories%' => $this->baseNamespace . $this->namespaceConfig . $this->subdirectory,
         ];
     }
 
     /**
-     * Takes care of the creation process of the contract file, if the file does exist
+     * Takes care of the creation process of the repository file, if the file does exist
      * it will abort the process and return false, else it will create it and return
-     * the full name space to it with it's name in an array.
      *
-     * @return mixed bool|array
+     * @return bool
      */
     public function create()
     {
-        // get the content of the stub file of contract
+        // get the content of the stub file of repository
         $this->extractStubContent($this->stub);
 
         // replacing each string that match a key in $replacements with the value of that key in $content
         $this->replaceContentParts($this->replacements);
 
-        // preparing repository contract (interface) class name
+        // preparing repository repository (interface) class name
         $this->createClassName($this->modelName);
 
-        // preparing the full path to the directory where the contracts will be stored
+        // preparing the full path to the directory where the repositories will be stored
         $this->generateDirectoryFullPath($this->directoryBasePath(), $this->pathConfig);
 
-        // checking that the directory of repository contracts doesn't exist
+        // checking that the directory of repository repositories doesn't exist
         if (!$this->directoryExists($this->directory)) {
             // if so, create the directory
             $this->createDirectory($this->directory);
         }
 
-        // preparing the full path to the repository contract file
+        // preparing the full path to the repository repository file
         $this->generateFileFullPath($this->directory, $this->className);
 
-        // checking that the repository contract file does not exist in the directory
+        // checking that the repository repository file does not exist in the directory
         if (!$this->fileExists($this->path)) {
             // creating th file
             $result = $this->createFile($this->path, $this->content);
@@ -91,7 +76,7 @@ class ContractCreator extends Creator
                 return false;
             }
             // else, return response
-            return $this->getReturnedData();
+            return true;
         }
 
         // if the file does exist, don't create the file & return false
@@ -101,7 +86,7 @@ class ContractCreator extends Creator
     /**
      * Completes the creation process when it was aborted due to the existence of the file.
      *
-     * @return array|bool
+     * @return bool
      */
     public function complete()
     {
@@ -112,20 +97,7 @@ class ContractCreator extends Creator
             return false;
         }
         // else, return response
-        return $this->getReturnedData();
-    }
-
-    /**
-     * Prepares the array that will be returned by the creation methods of the class.
-     *
-     * @return array
-     */
-    private function getReturnedData(): array
-    {
-        return [
-            $this->getClassFullNamespace($this->baseNamespace(), $this->className),
-            $this->className,
-        ];
+        return true;
     }
 
     public function getStub()
